Refactor VMState usage to VMStateInterpreter for improved clarity and consistency across execution contexts
